pyproject.toml
[tool.poetry]
name = "api-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
pymongo = "^4.1.1"
Flask = "1.1.2"
flask-restx = "^0.5.1"
fastapi = "^0.79.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
grpcio = "^1.48.0"
grpcio-tools = "^1.48.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
